requirements.txt
flask
gunicorn
pdfplumber
psycopg2-binary
python-dotenv
requests
pillow
